Documents all the public APIs for Nubody & markings code (#42857)
(cherry picked from commit 9c23b4a6d8d3188b8027ae5c0042a31455e31e03)
