+++ Content.Shared/Body/BodyComponent.cs
@@ -3,6 +3,11 @@ using Robust.Shared.GameStates;
 
 namespace Content.Shared.Body;
 
+/// <summary>
+/// Component on the entity that "has" a body, and that oversees entities with the <see cref="OrganComponent"/> inside it.
+/// </summary>
+/// <seealso cref="BodySystem" />
+/// <seealso cref="SharedVisualBodySystem" />
 [RegisterComponent, NetworkedComponent]
 [Access(typeof(BodySystem))]
 public sealed partial class BodyComponent : Component
