+++ Content.Shared/Body/BodySystem.Relay.cs
@@ -2,11 +2,13 @@ using Content.Shared.Body.Events;
 using Content.Shared.Gibbing;
 using Content.Shared.Humanoid;
 using Content.Shared.Medical;
+using JetBrains.Annotations;
 
 namespace Content.Shared.Body;
 
 public sealed partial class BodySystem
 {
+    // Refrain from adding an infinite block of relays here - consuming systems can use RelayEvent
     private void InitializeRelay()
     {
         SubscribeLocalEvent<BodyComponent, ApplyMetabolicMultiplierEvent>(RefRelayBodyEvent);
@@ -28,6 +30,13 @@ public sealed partial class BodySystem
         RelayEvent((uid, component), args);
     }
 
+    /// <summary>
+    /// Relays the given event to organs within a body.
+    /// </summary>
+    /// <param name="ent">The body to relay the event within</param>
+    /// <param name="args">The event to relay</param>
+    /// <typeparam name="T">The type of the event</typeparam>
+    [PublicAPI]
     public void RelayEvent<T>(Entity<BodyComponent> ent, ref T args) where T : struct
     {
         var ev = new BodyRelayedEvent<T>(ent, args);
@@ -38,6 +47,13 @@ public sealed partial class BodySystem
         args = ev.Args;
     }
 
+    /// <summary>
+    /// Relays the given event to organs within a body.
+    /// </summary>
+    /// <param name="ent">The body to relay the event within</param>
+    /// <param name="args">The event to relay</param>
+    /// <typeparam name="T">The type of the event</typeparam>
+    [PublicAPI]
     public void RelayEvent<T>(Entity<BodyComponent> ent, T args) where T : class
     {
         var ev = new BodyRelayedEvent<T>(ent, args);
@@ -49,7 +65,7 @@ public sealed partial class BodySystem
 }
 
 /// <summary>
-/// Event wrapper for relayed events.
+/// Event wrapper for events being relayed to organs within a body.
 /// </summary>
 [ByRefEvent]
 public record struct BodyRelayedEvent<TEvent>(Entity<BodyComponent> Body, TEvent Args);
