+++ Content.Shared/Body/VisualOrganComponent.cs
@@ -4,28 +4,35 @@ using Robust.Shared.Serialization;
 
 namespace Content.Shared.Body;
 
+/// <summary>
+/// Defines an organ that applies a sprite to the specified <see cref="Layer" /> within the body
+/// </summary>
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState(true)]
 [Access(typeof(SharedVisualBodySystem))]
 public sealed partial class VisualOrganComponent : Component
 {
     /// <summary>
-    /// The layer on the entity that this contributes to
+    /// The sprite layer on the entity that this contributes to
     /// </summary>
     [DataField(required: true)]
     public Enum Layer;
 
     /// <summary>
-    /// The data for the layer
+    /// The sprite data for the layer
     /// </summary>
     [DataField(required: true), AutoNetworkedField, AlwaysPushInheritance]
     public PrototypeLayerData Data;
 
     /// <summary>
-    /// When applying a profile, if the sex is present in this dictionary, overrides the state of the data.
+    /// When applying a profile, if the sex is present in this dictionary, overrides the state of the sprite data.
+    /// Used for e.g. male vs female torsoes.
     /// </summary>
     [DataField]
     public Dictionary<Sex, string>? SexStateOverrides;
 
+    /// <summary>
+    /// The current profile data of this organ, used for alternate sprite selection and colouration.
+    /// </summary>
     [DataField, AutoNetworkedField]
     public OrganProfileData Profile = new();
 }
